Corrected Exercises

+++ basics/src/main/java/be/ecam/basics/exercises/Counter.java
@@ -1,11 +1,24 @@
+// package be.ecam.basics.exercises;
+
+// public class Counter {
+//     public static byte count(int start, int steps) {
+//         byte c = (byte) start;
+//         for (int i = 0; i < steps; i++) {
+//             c++;
+//         }
+//         return c;
+//     }
+// }
+
 package be.ecam.basics.exercises;
 
 public class Counter {
-    public static byte count(int start, int steps) {
-        byte c = (byte) start;
+    public static int count(int start, int steps) {
+        int c = start;
         for (int i = 0; i < steps; i++) {
             c++;
         }
         return c;
     }
 }
+
